- Add controller, model, view for Users.
Add auto-add user on sign-in.
Add migration for Users.

- Add currentUser globally
- Small changes in layout
- Add creators migration
- Add limiting scope of data access for HR
- Add authorization on controllers initialization
- Make endpoints secure
- Update views for authorization.
- Fix unit tests

Related work items: #81

+++ HRPotter/Controllers/CompaniesController.cs
@@ -6,18 +6,25 @@ using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Authorization;
+using static HRPotter.Controllers.UsersController;
+
 
 namespace HRPotter.Controllers
 {
+    [ProducesResponseType(StatusCodes.Status403Forbidden)]
     [Route("[controller]")]
     [Authorize]
     public class CompaniesController : Controller
     {
         private readonly HRPotterContext _context;
 
-        public CompaniesController(HRPotterContext context)
+        public CompaniesController(HRPotterContext context, IHttpContextAccessor httpContextAccessor)
         {
             _context = context;
+            if (!IsAuthorized())
+            {
+                AuthorizeUser(_context, httpContextAccessor.HttpContext.User);
+            }
         }
 
         /// <summary>
@@ -30,6 +37,11 @@ namespace HRPotter.Controllers
         [HttpGet]
         public async Task<IActionResult> Index()
         {
+            if (HRPotterUser.Role != "Admin")
+            {
+                return Forbid();
+            }
+
             return View(await _context.Companies.ToListAsync());
         }
 
@@ -40,6 +52,11 @@ namespace HRPotter.Controllers
         [HttpGet]
         public async Task<IActionResult> GetCompaniesTable()
         {
+            if (HRPotterUser.Role != "Admin")
+            {
+                return Forbid();
+            }
+
             return PartialView("_CompaniesTable", await _context.Companies.ToListAsync());
         }
 
@@ -54,6 +71,11 @@ namespace HRPotter.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create(Company company)
         {
+            if (HRPotterUser.Role != "Admin")
+            {
+                return Forbid();
+            }
+
             if (ModelState.IsValid)
             {
                 _context.Add(company);
@@ -75,6 +97,11 @@ namespace HRPotter.Controllers
         [HttpGet]
         public async Task<IActionResult> Edit(int? id)
         {
+            if (HRPotterUser.Role != "Admin")
+            {
+                return Forbid();
+            }
+
             if (id == null)
             {
                 return NotFound();
@@ -100,6 +127,11 @@ namespace HRPotter.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(int id, [Bind("Id,Name")] Company company)
         {
+            if (HRPotterUser.Role != "Admin")
+            {
+                return Forbid();
+            }
+
             if (company == null || id != company.Id)
             {
                 return NotFound();
@@ -139,6 +171,11 @@ namespace HRPotter.Controllers
         [HttpGet]
         public async Task<IActionResult> GetDeleteModal(int? id)
         {
+            if (HRPotterUser.Role != "Admin")
+            {
+                return Forbid();
+            }
+
             if (id == null)
             {
                 return NotFound();
@@ -164,6 +201,11 @@ namespace HRPotter.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
+            if (HRPotterUser.Role != "Admin")
+            {
+                return Forbid();
+            }
+
             var company = await _context.Companies.FindAsync(id);
             _context.Companies.Remove(company);
             await _context.SaveChangesAsync();
